Rename and reorganise product module
