--- docs/code_ref/vis.rst
@@ -1,8 +1,0 @@
-STIXpy vis
-**************
-
-The ``vis`` submodule
-
-.. automodapi:: stixpy.vis
-
-.. automodapi:: stixpy.vis.map_reprojection
