+++ stixpy/product/__init__.py
@@ -1,0 +1,2 @@
+from stixpy.product.product_factory import Product
+from stixpy.product.sources import *
