+++ stixpy/product/product.py
