+++ stixpy/product/sources/__init__.py
@@ -1,0 +1,3 @@
+from stixpy.product.sources.housekeeping import *
+from stixpy.product.sources.quicklook import *
+from stixpy.product.sources.science import *
