--- stixpy/products/__init__.py
@@ -1,3 +1,0 @@
-from stixpy.products.product_factory import Product
-from stixpy.products.science import *
-from stixpy.products.quicklook import *
